Organize About activity and licences fragment
