+++ app/src/main/res/layout/activity_settings.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.khmelenko.lab.varis.activity.AboutActivity">
+    tools:context="com.khmelenko.lab.varis.about.AboutActivity">
 
     <include
         android:id="@+id/toolbar"
